fix: deep scan — signed/unsigned warnings, loop logic, dispose guard, typos, missing #include
Geometry.cpp:
- Loop variables changed from int to size_t to fix signed/unsigned comparison warnings

Geometry.h:
- Added missing #include <glm/glm.hpp> (needed for glm::mat4 in applyTransformationMatrix)
- Added virtual destructor to base class (Rule of 5 / polymorphic delete safety)

FirstPersonCamera.h:
- orientation changed from glm::vec3 to glm::dvec3 (consistent with double-precision lookSpeed/delta math)

ChunkNode.cpp:
- dispose(): added guard 'if (neighbors[i] != nullptr)' for ALL neighbors before accessing ->state,
  not just those with state > DATA — prevents null-deref crash when a neighbor slot was never filled
- getNodes(): loop variable changed to size_t
- getGeometries(): loop variable changed to size_t
- searchNode(): loop variables changed to size_t
- Typo fix in comment: 'it it does not' -> 'if it does not'

VkCraft.h:
- Typo fix: 'Chose swap chain' -> 'Choose swap chain' (chooseSwapPresentMode doc comment)"

+++ src/vkCraft/FirstPersonCamera.h
@@ -17,7 +17,9 @@ class FirstPersonCamera : public Camera
 public:
 	glm::dvec2 last = { 0.0, 0.0 };
 	glm::dvec2 delta = { 0.0, 0.0 };
-	glm::vec3 orientation = { 0.0f, 0.0f, 0.0f };
+
+	// dvec3 matches the double-precision arithmetic used in update() for lookSpeed/delta
+	glm::dvec3 orientation = { 0.0, 0.0, 0.0 };
 
 	/**
 	 * Tracks whether last cursor position has been initialized.
