+++ src/vkCraft/Geometry.cpp
@@ -9,7 +9,7 @@ bool Geometry::hasBuffers()
 
 void Geometry::applyTransformationMatrix(glm::mat4 *matrix)
 {
-	for (int i = 0; i < vertices.size(); i++)
+	for (size_t i = 0; i < vertices.size(); i++)
 	{
 		glm::vec4 result = *matrix * glm::vec4(vertices[i].pos, 1.0f);
 			
@@ -21,16 +21,16 @@ void Geometry::applyTransformationMatrix(glm::mat4 *matrix)
 
 void Geometry::merge(Geometry *geometry)
 {
-	int initialSize = vertices.size();
+	size_t initialSize = vertices.size();
 
-	for (int i = 0; i < geometry->vertices.size(); i++)
+	for (size_t i = 0; i < geometry->vertices.size(); i++)
 	{
 		vertices.push_back(geometry->vertices[i]);
 	}
 
-	for (int i = 0; i < geometry->indices.size(); i++)
+	for (size_t i = 0; i < geometry->indices.size(); i++)
 	{
-		indices.push_back(geometry->indices[i] + initialSize);
+		indices.push_back(geometry->indices[i] + static_cast<uint32_t>(initialSize));
 	}
 }
